fix(victron): use int32 for /Hub4/L1/AcPowerSetpoint + fix fallback empty-stdout check (#186)
- vebus AcPowerSetpoint is int32, not int16: every SetValue was being
  rejected by GX. Other dbus_set call sites already use int32 (settings,
  Hub4Mode); only set_grid_setpoint had the wrong type.
- _safe_subprocess treated empty stdout (dbus-send SetValue success has
  no stdout) as failure, masking successful writes as failed.

Fixes the publish wedge after ./deploy.sh: control loop now actually
writes the setpoint. Native dbus_fast SetValue still has a self-call
deadlock concern (returns None when called from the loop thread) — the
subprocess fallback path now correctly reports success.

Tests: test_set_grid_setpoint updated to assert int32.

+++ tests/test_dbus_native.py
@@ -208,7 +208,7 @@ class TestVictronDBusIntegration:
 
         assert v.set_grid_setpoint(-615)
         fake.set_value.assert_called_once_with(
-            "com.victronenergy.vebus.ttyUSB2", "/Hub4/L1/AcPowerSetpoint", -615, "int16"
+            "com.victronenergy.vebus.ttyUSB2", "/Hub4/L1/AcPowerSetpoint", -615, "int32"
         )
         mock_run.assert_not_called()
         assert v._consecutive_errors == 0
